relacion licencias con tareas

+++ src/Sofco.Core/DAL/WorkTimeManagement/IWorkTimeRepository.cs
@@ -12,7 +12,7 @@ namespace Sofco.Core.DAL.WorkTimeManagement
 
         IList<WorkTime> Search(WorktimeHoursApprovedParams model);
 
-        int GetTotalHoursByDate(DateTime date, int currentUserId);
+        decimal GetTotalHoursByDate(DateTime date, int currentUserId);
 
         IList<WorkTime> SearchPending(WorktimeHoursPendingParams model, bool isManagerOrDirector, int currentUserId);
