+++ src/Sofco.Core/Models/WorkTimeManagement/WorkTimeResumeModel.cs
@@ -2,18 +2,18 @@
 {
     public class WorkTimeResumeModel
     {
-        public int HoursApproved { get; set; }
+        public decimal HoursApproved { get; set; }
 
-        public int HoursRejected { get; set; }
+        public decimal HoursRejected { get; set; }
 
-        public int HoursPending { get; set; }
+        public decimal HoursPending { get; set; }
 
-        public int HoursPendingApproved { get; set; }
+        public decimal HoursPendingApproved { get; set; }
 
-        public int Total => HoursApproved + HoursRejected + HoursPending + HoursPendingApproved;
+        public decimal Total => HoursApproved + HoursRejected + HoursPending + HoursPendingApproved;
 
-        public int BusinessHours { get; set; }
+        public decimal BusinessHours { get; set; }
 
-        public int HoursUntilToday { get; set; }
+        public decimal HoursUntilToday { get; set; }
     }
 }
