+++ src/Sofco.Model/Models/WorkTimeManagement/WorkTime.cs
@@ -21,7 +21,7 @@ namespace Sofco.Model.Models.WorkTimeManagement
 
         public DateTime Date { get; set; }
 
-        public int Hours { get; set; }
+        public decimal Hours { get; set; }
 
         public string Source { get; set; }
